Allocate sound samples on the stack + refactoring

+++ mars_newrb.h
@@ -29,7 +29,7 @@
 typedef struct {
     volatile int readpos, writepos;
     volatile int n_writepos; // the value writepos is going to take after the next commit
-    volatile int maxreadpos;
+    volatile uint16_t maxreadpos;
     volatile char ropen, wopen;
     volatile char lock;
     char nolock;
